# openapi.yaml
openapi: 3.0.0
paths:
...
  /residents/{residentUid}/food-records:
    get:
      operationId: getFoodRecords
      parameters:
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema:
            type: string
      responses:
        "200":
          description: 食事記録一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodRecordListResponseDto"
      security:
        - JWT-auth: []
      summary: 利用者の食事記録一覧を取得
      tags: &a5
        - food-records
    post:
      operationId: createFoodRecord
      parameters:
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodRecordCreateInputDto"
      responses:
        "201":
          description: 食事記録の作成に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodRecordDto"
      security:
        - JWT-auth: []
      summary: 食事記録を作成
      tags: *a5
  /residents/{residentUid}/food-records/daily:
    get:
      operationId: getDailyFoodRecords
      parameters:
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema:
            type: string
        - name: startDate
          required: false
          in: query
          description: 開始日（YYYY-MM-DD形式）。指定しない場合は過去30日間
          schema:
            type: string
        - name: endDate
          required: false
          in: query
          description: 終了日（YYYY-MM-DD形式）。指定しない場合は現在日
          schema:
            type: string
      responses:
        "200":
          description: 日別食事記録一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyFoodRecordsListResponseDto"
      security:
        - JWT-auth: []
      summary: 利用者の日別食事記録一覧を取得
      tags: *a5
  /residents/{residentUid}/food-records/{uid}:
    patch:
      operationId: updateFoodRecord
      parameters:
        - name: uid
          required: true
          in: path
          description: 食事記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodRecordUpdateInputDto"
      responses:
        "200":
          description: 食事記録の更新に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodRecordDto"
      security:
        - JWT-auth: []
      summary: 食事記録を更新
      tags: *a5
    delete:
      operationId: deleteFoodRecord
      parameters:
        - name: uid
          required: true
          in: path
          description: 食事記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 食事記録の削除に成功
      security:
        - JWT-auth: []
      summary: 食事記録を削除
      tags: *a5
  /residents/{residentUid}/food-records/{uid}/transcription:
    get:
      operationId: getFoodRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 食事記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 食事記録の文字起こし取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: 食事記録の文字起こしを取得
      tags: *a5
    patch:
      operationId: appendFoodRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 食事記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: 食事記録の文字起こし追記に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: 食事記録の文字起こしを追記
      tags: *a5
    put:
      operationId: updateFoodRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 食事記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: 食事記録の文字起こし置換に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: 食事記録の文字起こしを置換
      tags: *a5
    delete:
      operationId: deleteFoodRecordTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: 食事記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 食事記録の文字起こし削除に成功
      security:
        - JWT-auth: []
      summary: 食事記録の文字起こしを削除
      tags: *a5
  /residents/{residentUid}/food-records/{uid}/extract:
    post:
      operationId: extractFoodRecord
      parameters:
        - name: uid
          required: true
          in: path
          description: 食事記録UID
          schema:
            type: string
        - name: residentUid
          required: true
          in: path
          description: 利用者UID
          schema: {}
      responses:
        "200":
          description: 食事記録からの情報抽出に成功
          content:
            application/json:
              schema:
                type: string
      security:
        - JWT-auth: []
      summary: 食事記録から情報を抽出
      tags: *a5
...
  /assessments/{uid}/transcription:
    get:
      operationId: getTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      responses:
        "200":
          description: アセスメントの文字起こし取得に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを取得
      tags: *a10
    patch:
      operationId: appendTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: アセスメントの文字起こし追記に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを追記
      tags: *a10
    put:
      operationId: updateTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionInputDto"
      responses:
        "200":
          description: アセスメントの文字起こし置換に成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionDto"
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを置換
      tags: *a10
    delete:
      operationId: deleteTranscription
      parameters:
        - name: uid
          required: true
          in: path
          description: アセスメントUID
          schema:
            type: string
      responses:
        "200":
          description: アセスメントの文字起こし削除に成功
      security:
        - JWT-auth: []
      summary: アセスメントの文字起こしを削除
      tags: *a10
...
info:
  title: Nagara Care API
  description: ""
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
...
    FoodRecordDto:
      type: object
      properties:
        uid:
          type: string
          description: UID
        tenantUid:
          type: string
          description: テナントUID
        caregiverUid:
          type: string
          description: 介護者UID
        residentUid:
          type: string
          description: 利用者UID
        recordedAt:
          format: date-time
          type: string
          description: 記録時刻
        notes:
          type: string
          description: メモ
        mealTime:
          type: string
          description: 食事の時間帯
          enum:
            - BREAKFAST
            - LUNCH
            - DINNER
        mainCoursePercentage:
          type: number
          description: 主食の摂取率（%）
        sideDishPercentage:
          type: number
          description: 副食の摂取率（%）
        soupPercentage:
          type: number
          description: 汁物の摂取率（%）
        beverageType:
          type: string
          description: 飲み物の種類
          enum:
            - WATER
            - TEA
            - OTHER
        beverageVolume:
          type: number
          description: 飲み物の摂取量（ml）
      required:
        - uid
        - tenantUid
        - caregiverUid
        - residentUid
        - recordedAt
        - notes
        - mealTime
        - mainCoursePercentage
        - sideDishPercentage
        - soupPercentage
        - beverageType
        - beverageVolume
    FoodRecordListResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FoodRecordDto"
        total:
          type: number
          description: 総件数
      required:
        - items
        - total
    DailyFoodRecordsDto:
      type: object
      properties:
        date:
          type: string
          description: 日付
          example: 2024-03-20
        breakfast:
          description: 朝食の記録
          allOf:
            - $ref: "#/components/schemas/FoodRecordDto"
        lunch:
          description: 昼食の記録
          allOf:
            - $ref: "#/components/schemas/FoodRecordDto"
        dinner:
          description: 夕食の記録
          allOf:
            - $ref: "#/components/schemas/FoodRecordDto"
      required:
        - date
    DailyFoodRecordsListResponseDto:
      type: object
      properties:
        items:
          description: 日別食事記録一覧
          type: array
          items:
            $ref: "#/components/schemas/DailyFoodRecordsDto"
        total:
          type: number
          description: 総日数
          example: 10
      required:
        - items
        - total
    FoodRecordCreateInputDto:
      type: object
      properties:
        recordedAt:
          format: date-time
          type: string
          description: 記録時刻
          example: 2024-03-20T12:00:00Z
        notes:
          type: string
          description: メモ
        mealTime:
          type: string
          description: 食事の時間帯
          enum:
            - BREAKFAST
            - LUNCH
            - DINNER
          example: LUNCH
        mainCoursePercentage:
          type: number
          description: 主食の摂取率（%）
          minimum: 0
          maximum: 100
          example: 80
        sideDishPercentage:
          type: number
          description: 副食の摂取率（%）
          minimum: 0
          maximum: 100
          example: 70
        soupPercentage:
          type: number
          description: 汁物の摂取率（%）
          minimum: 0
          maximum: 100
          example: 90
        beverageType:
          type: string
          description: 飲み物の種類
          enum:
            - WATER
            - TEA
            - OTHER
          example: WATER
        beverageVolume:
          type: number
          description: 飲み物の摂取量（ml）
          minimum: 0
          example: 200
      required:
        - recordedAt
        - mealTime
        - mainCoursePercentage
        - sideDishPercentage
        - soupPercentage
        - beverageType
        - beverageVolume
    FoodRecordUpdateInputDto:
      type: object
      properties:
        recordedAt:
          format: date-time
          type: string
          description: 記録時刻
          example: 2024-03-20T12:00:00Z
        notes:
          type: string
          description: メモ
        mealTime:
          type: string
          description: 食事の時間帯
          enum:
            - BREAKFAST
            - LUNCH
            - DINNER
          example: LUNCH
        mainCoursePercentage:
          type: number
          description: 主食の摂取率（%）
          minimum: 0
          maximum: 100
          example: 80
        sideDishPercentage:
          type: number
          description: 副食の摂取率（%）
          minimum: 0
          maximum: 100
          example: 70
        soupPercentage:
          type: number
          description: 汁物の摂取率（%）
          minimum: 0
          maximum: 100
          example: 90
        beverageType:
          type: string
          description: 飲み物の種類
          enum:
            - WATER
            - TEA
            - OTHER
          example: WATER
        beverageVolume:
          type: number
          description: 飲み物の摂取量（ml）
          minimum: 0
          example: 200
      required:
        - recordedAt
        - mealTime
        - mainCoursePercentage
        - sideDishPercentage
        - soupPercentage
        - beverageType
        - beverageVolume
    TranscriptionDto:
      type: object
      properties:
        transcription:
          type: string
          description: 文字起こしテキスト
      required:
        - transcription
    TranscriptionInputDto:
      type: object
      properties:
        transcription:
          type: string
          description: 文字起こしテキスト
      required:
        - transcription
...